% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-funcs.R
\docType{methods}
\name{freq}
\alias{freq}
\title{freq}
\usage{
freq(x, y, x.n = 11, y.n = x.n, na.rm = FALSE)
}
\arguments{
\item{x}{a vector holding a time series}

\item{y}{a vector holding a time series}

\item{x.n}{a numeric vector giving number of bins}

\item{y.n}{a numeric vector giving number of bins}

\item{na.rm}{logical; if true, any NA and NaN's are removed from x before calculations}
}
\value{
a \code{data.frame} frequency by bin
}
\description{
Calculates the frequency of an obervation in a 2D cell
}
\examples{
\dontrun{
   y=rnorm(20)
   x=rnorm(20)
   freq(x,y)}
}
