\name{theme_flr}
\alias{theme_flr}
\title{A personalised theme with white background and no gridlines.}
\usage{
  theme_flr(base_size=16,font=NA,face='plain', 
            backgroundColor='white', panelColor=backgroundColor, 
            axisColor='#999999', gridColor= backgroundColor, textColor='black',
            axis.text=1.0,angle.x=0, angle.y=-90,...)
}
\arguments{
  \item{base_size}{base font size}
  \item{font}{base font family}
  \item{face}{b} 
  \item{backgroundColor}{colour of the plot background} 
  \item{panelColor}{colour of a panel} 
  \item{axisColor}{colour of axes} 
  \item{gridColor}{colour of grid} 
  \item{textColor}{colour of text}
  \item{axis.text}{axis text}
  \item{angle.x}{angle for text on x axis} 
  \item{angle.y}{angle for text on y axis}
  \item{...}{other args}
                     
}
\description{
  A personalised theme for flr, making it easy to personalise plots.
}
\keyword{theme}

\examples{

## Not Run
## End(Not run)

}
\seealso{

\code{\link{multiplot}}
}


