% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{hinge}
\alias{hinge}
\title{incr}
\usage{
hinge(s, v, rec)
}
\arguments{
\item{s}{steepness}

\item{v}{virgin biomass}

\item{rec}{a vector holding constant recruitment level}
}
\value{
a \code{numeric} giving break point
}
\description{
Break point for segmented regression stock recruitment relationship
for a given steepness and virgin biomass
}
\examples{
\dontrun{
   hinge(.7,1000,2000)}
}

