% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-funcs.R
\docType{methods}
\name{iav}
\alias{iav}
\title{iav}
\usage{
iav(x)
}
\arguments{
\item{x}{a vector holding a time series}
}
\value{
a \code{vector} with the inter-annual variation each time step
}
\description{
Calculates the inter-annual variation in a time series, i.e. \code{(x[t+1]-x[t])/x[t]}
Used to show how variable a quantity like yield 
is under different management strategies within a Management Strategy Evaluation.
}
\examples{
\dontrun{
   x=rnorm(2)
   iav(x)
   ## inter-annual average variation
   mean(iav(x),na.rm=T)}
}

