% $Id: data.Rd 432 2010-02-05 22:19:53Z imosqueira $
\name{data}
\alias{yft}
\title{Example Datasets}
\description{
\itemize{
     \item{yft}{ A data.frame with historical estimates of stock biomass and exploitation level relative  to $B_{MSY}$ and $F_{MSY}$ for the 2016 stock assessment of Atlantic yellowfin tuna.}
	Datasets can be loaded by issuing the \code{data} command, e.g. \code{data(sims)}.
}
\references{
}
\keyword{datasets}
