% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-2box.R
\docType{methods}
<<<<<<< HEAD:man/kobe2box.Rd
\name{kobe2box}
\alias{kobe2box}
\alias{kobe2box,character-method}
\alias{kobe2box-method}
\title{kobe2box}
=======
\name{kobe.2box}
\alias{kobe.2box}
\alias{kobe.2box-method}
\alias{kobe.2box,character-method}
\title{kobe.2box}
>>>>>>> 6ca268c754db91b327ca33d14ae9c3ceafd60c08:man/kobe.2box.Rd
\usage{
\S4method{kobe2box}{character}(path, proxy = c("fmsy", "fmax", "f0.1", "f20",
  "f30", "f40", "f90max", "f75max")[3], what = c("sims", "trks", "pts",
  "smry", "wrms")[1], prob = c(0.75, 0.5, 0.25), year = NULL, nwrms = 10)
}
\arguments{
\item{path}{\code{character} directory or directories with target files}

\item{proxy}{\code{character} proxy for msy "f0.1" by default otherwise one of c("fmsy","fmax","f0.1","f20","f30","f40","f90max","f75max")}

\item{what}{\code{character} output data.frame all "sims" by default or one of c("sims","trks","pts","smry","wrms")[1]}

\item{prob}{\code{numeric} with probabilities, by default c(0.75,0.5,0.25)}

\item{year}{\code{numeric} for year when slecting points}

\item{nwrms}{\code{numeric} number of worms to select by default 10}
}
\value{
a \code{data.frame} or a list of \code{data.frame} with either all the simulations
from PRO2BOX, i.e. for both historical assessment results and projections. These can either
be in the form of summaries or all reults
}
\description{
Creates summary data.frame for presenting advice using VPA2Box projections
}
\examples{
\dontrun{
sims=kobe2box("vpa")
} 
}
