% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{kobePhase}
\alias{kobePhase}
\alias{kobePhase,data.frame-method}
\alias{kobePhase,missing-method}
\alias{kobePhase-method}
\title{kobePhase}
\usage{
\S4method{kobePhase}{missing}(object, xlim = c(0, 2), ylim = xlim,
  quadcol = c("yellow", "green", "red"))
}
\arguments{
\item{object}{an object of class \code{missing,data.frame,...}}

\item{xlim}{a numeric vector with x-axis limits, by default is c(0.2)}

\item{ylim}{a numeric vector with y-axis limits, by default is the same as xlim}
}
\description{
Creates The Kobe Phase Plot
}
\examples{
\dontrun{kobePhase()}
}

