% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-sam.R, R/kobe-sam-tmb.R
\name{kobeSam}
\alias{kobeSam}
\alias{kobeSam-method}
\alias{kobeSam,character-method}
\alias{kobeSam}
\alias{kobeSam-method}
\alias{kobeSam,character-method}
\title{kobeSam}
\arguments{
\item{path}{\code{character} name of target files}

\item{what}{\code{character} output data.frame all "sims" by default or one of c("sims","trks","pts","smry","wrms",ellipse)[1]}

\item{prob}{\code{numeric} probabilities if trks output, by default c(0.75,0.5,0.25)}

\item{year}{\code{numeric} for year when slecting points}

\item{nits}{\code{numeric} number of iterations for simulations}

\item{nwrms}{\code{numeric} number of worms to select by default 10}

\item{bmsy}{\code{numeric} value for biomass reference point}

\item{fmsy}{\code{numeric} value for F reference point}

\item{path}{\code{character} name of target files}

\item{what}{\code{character} output data.frame all "sims" by default or one of c("sims","trks","pts","smry","wrms",ellipse)[1]}

\item{prob}{\code{numeric} probabilities if trks output, by default c(0.75,0.5,0.25)}

\item{year}{\code{numeric} for year when slecting points}

\item{nits}{\code{numeric} number of iterations for simulations}

\item{nwrms}{\code{numeric} number of worms to select by default 10}

\item{bmsy}{\code{numeric} value for biomass reference point}

\item{fmsy}{\code{numeric} value for F reference point}
}
\value{
a \code{data.frame} or a list of \code{data.frame} with historical simulations
from SAM, i.e. for both historical assessment results and projections. These can either
be in the form of summaries or all reults

a \code{data.frame} or a list of \code{data.frame} with historical simulations
from SAM, i.e. for both historical assessment results and projections. These can either
be in the form of summaries or all reults
}
\description{
Creates summary data.frame for presenting advice using SAM projections

Creates summary data.frame for presenting advice using SAM projections
}
\examples{
\dontrun{
sims=kobeSam("vpa")
} 
\dontrun{
sims=kobeSam("vpa")
} 
}
