% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{kobeSmry}
\alias{kobeSmry}
\alias{kobeSmry,data.frame,missing-method}
\alias{kobeSmry,numeric,numeric-method}
\alias{kobeSmry-method}
\title{Creates Summaries i.e. probability of being in the green quadrant}
\usage{
\S4method{kobeSmry}{numeric,numeric}(stock, harvest, onlyColours = FALSE)
}
\arguments{
\item{stock}{an object of class \code{vector,data.frame,...}}

\item{harvest}{an object of class \code{vector,data.frame,...}}

\item{onlyColours}{\code{logical} to report only quadrant colours}
}
\description{
Creates tracks for plotting in the Kobe Phase Plot
}
\examples{
\dontrun{kobeSmry()}
}

