% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{kobeTrks}
\alias{kobeTrks}
\alias{kobeTrks,data.frame,missing-method}
\alias{kobeTrks,numeric,numeric-method}
\alias{kobeTrks-method}
\title{Creates tracks for plotting on phase plot}
\usage{
\S4method{kobeTrks}{numeric,numeric}(stock, harvest, prob = c(0.25, 0.5,
  0.75), na.rm = FALSE)
}
\arguments{
\item{stock}{an object of class \code{vector,data.frame,...}}

\item{harvest}{an object of class \code{vector,data.frame,...}}

\item{prob}{a numeric vector with probabilities}

\item{na.rm}{logical}
}
\description{
Creates tracks for plotting in the Kobe Phase Plot
}
\examples{
\dontrun{kobeTrks()}
}

