% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-multiplot.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(..., plotlist = NULL, cols = 1)
}
\arguments{
\item{...}{\code{ggplot} objects}

\item{plotlist}{a list \code{ggplot} objects}

\item{cols}{number of columns in plot grid, by default=1}
}
\value{
an \code{ggplot2} object
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
- cols:   Number of columns in layout
- layout: A matrix specifying the layout. If present, 'cols' is ignored.
If the layout is something like matrix(c(1,2,3,3), nrow=2, by.row=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and

borrowed with thanks from
http://wiki.stdout.org/rcookbook/Graphs/Multiple%20graphs%20on%20one%20page%20(ggplot2)/
}
\examples{
\dontrun{
multiplot(plot(1),plot(2))
} 
}

