% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-shade.R
\docType{methods}
\name{shade}
\alias{shade}
\alias{shade-method}
\alias{shade,numeric-method}
\alias{shade,matrix-method}
\alias{shade,data.frame-method}
\title{Creates Kobe II Strategy Matrices}
\usage{
\S4method{shade}{numeric}(object, breaks = c(-0.1, 50, 60, 70, 80, 90, 100),
  shades = c("\\\\cellcolor{gray100}", "\\\\cellcolor{gray90}",
  "\\\\cellcolor{gray80}", "\\\\cellcolor{gray70}", "\\\\cellcolor{gray60}",
  "\\\\cellcolor{gray50}"), pct = "\\\\\%", ...)
}
\arguments{
\item{object}{an object of class \code{vector,data.frame,...}}

\item{breaks}{vector of breaks}

\item{shades}{vector of colours for shading}

\item{pct}{adds percent sign}

\item{...}{other arguments}
}
\description{
Creates latex tables with shading for inclusion in latex docs.
}
