% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-smry.R
\docType{methods}
\name{smry}
\alias{smry}
\alias{smry-method}
\alias{smry,numeric,numeric-method}
\alias{smry,data.frame,missing-method}
\title{Creates Summaries i.e. probability of being in the green quadrant}
\usage{
\S4method{smry}{numeric,numeric}(stock, harvest, onlyColours = FALSE)
}
\arguments{
\item{stock}{an object of class \code{vector,data.frame,...}}

\item{harvest}{an object of class \code{vector,data.frame,...}}

\item{onlyColours}{\code{logical} to report only quadrant colours}
}
\description{
Creates tracks for plotting in the Kobe Phase Plot
}
\examples{
\dontrun{smry()}

}
