% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-trks.R
\docType{methods}
\name{trks}
\alias{trks}
\alias{trks-method}
\alias{trks,numeric,numeric-method}
\alias{trks,data.frame,missing-method}
\title{Creates tracks for plotting on phase plot}
\usage{
\S4method{trks}{numeric,numeric}(stock, harvest, prob = c(0.25, 0.5,
  0.75), na.rm = FALSE)
}
\arguments{
\item{stock}{an object of class \code{vector,data.frame,...}}

\item{harvest}{an object of class \code{vector,data.frame,...}}

\item{prob}{a numeric vector with probabilities}

\item{na.rm}{logical}
}
\description{
Creates tracks for plotting in the Kobe Phase Plot
}
\examples{
\dontrun{trks()}

}
