% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-data.R
\docType{data}
\name{yft}
\alias{yft}
\title{yft}
\format{A \code{data.frame}
\describe{
  \item{method}{stock assessment method \code{factor}}
  \item{scenario}{stock assessment scenario \code{factor}}
  \item{iter}{bootstrap simulation \code{numeric}}
  \item{year}{historical assessment year \code{numeric}}
  \item{stock}{biomaass relative to BMSY \code{numeric}}   
  \item{harvest}{exploitation relative to FMSY \code{numeric}}
}}
\description{
An example dataset for Atlantic yellowfin
}
\references{
\url{http://http://iccat.int/Documents/CVSP/CV070_2014/n_4/CV070041365.pdf/}
}
\author{
Laurence Kell \email{laurie@kell.es}
}
\keyword{data}
