% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dRate}
\alias{dRate}
\title{dRate}
\usage{
dRate(x, r, wtAv = FALSE)
}
\arguments{
\item{x}{a vector holding a time series}

\item{r}{a numeric discount rate}

\item{wtAv}{a logical, FALSE by default, if TRUE then uses discount rate to calculate a weighted average}
}
\value{
net present value
}
\description{
Discount rate reflects the risk (the higher the risk the higher the discount rate). Is used
to discount all forecast future cash flows to calculate a present value:
}
\examples{
\dontrun{
   x=rnorm(20)
   dRate(x,0.05)}
}

