% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-funcs.R
\docType{methods}
\name{incr}
\alias{incr}
\title{incr}
\usage{
incr(x)
}
\arguments{
\item{x}{a vector holding a time series}
}
\value{
a \code{logical} indicating an increase
}
\description{
Is a quantity increasing from 1 time step to another \code{(x[t+1]-x[t])>0}
}
\examples{
\dontrun{
   x=rnorm(2)
   incr(x)}
}
