% $Id: data.Rd 432 2010-02-05 22:19:53Z imosqueira $
\name{data}
\alias{sims}
\alias{assmt}
\alias{prj1}
\alias{prj}
\title{Example Datasets}
\description{
\itemize{
     \item{sims}{ A data.frame with histrorical and projected stock biomass and harvest rate by bootstrap iteration  year and assessment run, derived from ASPIC}
     \item{assmt}{A data.frame with histrorical stock biomass and harvest rate by bootstrap iteration year derived from ASPIC}
     \item{prj1}{ A data.frame with histrorical and projected stock biomass and harvest rate by bootstrap iteration year derived from ASPIC}
     \item{prj}{  A list of data.frames with example summaries of assessment and projection}
}
	Datasets can be loaded by issuing the \code{data} command, e.g. \code{data(sims)}.
}
\references{
tRFMO.
}
\keyword{datasets}
