% $Id: data.Rd 432 2010-02-05 22:19:53Z imosqueira $
\name{data}
\alias{sims}
\title{Example Datasets}
\description{
\itemize{
     \item{sims}{A data.frame with stock biomass and harvest rate by bootstrap iteration  year and assessment run, derived from ASPIC}
}
	Datasets can be loaded by issuing the \code{data} command, e.g. \code{data(sims)}.
}
\references{
tRFMO.
}
\keyword{datasets}
