% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe-phase.R
\docType{methods}
\name{kobePhase}
\alias{kobePhase}
\alias{kobePhase-method}
\alias{kobePhase,missing-method}
\alias{kobePhase,data.frame-method}
\title{kobePhase}
\usage{
\S4method{kobePhase}{missing}(object, xlim = c(0, 2), ylim = xlim,
  quadcol = c("red", "green", "yellow", "gold"), bref = 1, fref = 1)
}
\arguments{
\item{object}{an object of class \code{missing,data.frame,...}}

\item{xlim}{a numeric vector with x-axis limits, by default is c(0.2)}

\item{ylim}{a numeric vector with y-axis limits, by default is the same as xlim}

\item{quadcol}{colours for the quadrants}

\item{bref}{vertical seperation between quadants}

\item{fref}{horizontal seperation between quadants}
}
\description{
Creates The Kobe Phase Plot
}
\examples{
\dontrun{kobePhase()}
}
